+++ tooling/eslint/CHANGELOG.md
@@ -10,4 +10,4 @@
 
 ### Patch Changes
 
-- c81587f: use-search-params-state mvp release
+- c81587f: sp-hooks mvp release