+++ tooling/eslint/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sp-hooks/eslint-config
 
+## 0.3.0
+
+### Minor Changes
+
+- 79e7170: Releasing v0.1.0
+- 79e7170: Rebrand to sp-hooks + more functional approach.
+
 ## 0.2.2
 
 ### Patch Changes